% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_meta_element.R
\name{add_meta_element}
\alias{add_meta_element}
\title{Add a meta element to the HTML head}
\usage{
add_meta_element(x, name = NULL, content = NULL, charset = NULL,
  http_equiv = NULL)
}
\description{
Add metadata to the HTML document with a
\code{<meta>} element.
}
