% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_title_element.R
\name{add_title_element}
\alias{add_title_element}
\title{Add the title element to the HTML head}
\usage{
add_title_element(x, title)
}
\arguments{
\item{title}{the title of the HTML document.}
}
\description{
Add a title to the HTML document.
}
