% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperMCI.R
\name{AO}
\alias{AO}
\title{AO intervals}
\usage{
AO(N, n, a)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{a}{alpha}
}
\value{
AO intervals
}
\description{
Create alpha optimal acceptance intervals
}
\examples{
AO_interval <- AO(100,10,0.05);
}
