% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperMCI.R
\name{CI_interval}
\alias{CI_interval}
\title{CI intervals}
\usage{
CI_interval(N, n, a)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{a}{alpha}
}
\value{
CI intervals
}
\description{
Create confidence intervals by inverting symmetric alpha optimal acceptance intervals
}
\examples{
CI_interval_table <- CI_interval(100,10,0.05);
}
