% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf_int.R
\name{conf_int}
\alias{conf_int}
\title{This function takes in a vector and outputs the 95\% confidence interval of the sample mean.}
\usage{
conf_int(data)
}
\arguments{
\item{data}{A 1D vector of numbers. Non-numerical values result in an error.}
}
\value{
A 2-element 1D vector indicating the start and end of the 95\% confidence interval.
}
\description{
This function takes in a vector and outputs the 95\% confidence interval of the sample mean.
}
