% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperMCI.R
\name{cover_plot}
\alias{cover_plot}
\title{coverage probability plot}
\usage{
cover_plot(N, n, a, CI_table)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{a}{level}

\item{CItable}{the set of confidence intervals}
}
\value{
coverage plot
}
\description{
plot the set of coverage probabilities
}
