% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperMCI.R
\name{coverprob}
\alias{coverprob}
\title{coverage probabilities}
\usage{
coverprob(N, n, CItable)
}
\arguments{
\item{N}{population size}

\item{n}{sample size}

\item{CItable}{the set of confidence intervals}
}
\value{
coverage
}
\description{
Calculate the set of coverage probabilities
}
