% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_css_rules.R
\name{create_css_rules}
\alias{create_css_rules}
\title{Create a set of CSS rules}
\usage{
create_css_rules(selectors, rules = NULL)
}
\arguments{
\item{selectors}{the selector or set of selectors
to use for the CSS rule set.}

\item{rules}{a list of rules to use in the rule
set's declaration block.}
}
\description{
Create a set of CSS rules that can be added
to an HTML document or written to a file.
}
