% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{db.get}
\alias{db.get}
\title{Fetch gene sets}
\usage{
db.get(symbol)
}
\arguments{
\item{symbol}{A symbol corresponding to a gene set}
}
\value{
A list of gene sets
}
\description{
Fetch gene sets
}
\examples{
REACTOME <- db.get("C2.CP.REACTOME")

}
