% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{db_get}
\alias{db_get}
\title{Fetch gene sets}
\usage{
db_get(symbol)
}
\arguments{
\item{symbol}{A symbol corresponding to a gene set (hypeR comes with C2.CP.BIOCARTA, C2.CP.KEGG, and C2.CP.REACTOME, see vignette to download more)}
}
\value{
A list of gene sets
}
\description{
Fetch gene sets
}
\examples{
REACTOME <- db_get("C2.CP.REACTOME")

}
