% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.check_overlap}
\alias{.check_overlap}
\title{Check overlap of signature across genesets}
\usage{
.check_overlap(signature, genesets, threshold = 0.05)
}
\arguments{
\item{signature}{A vector of symbols}

\item{genesets}{A list of genesets}

\item{threshold}{Minimum percent overlap}
}
\value{
Percent overlap
}
\description{
Check overlap of signature across genesets
}
\keyword{internal}
