% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_dots.R
\name{.dots_multi_plot}
\alias{.dots_multi_plot}
\title{Plot top enriched genesets across multiple signatures}
\usage{
.dots_multi_plot(
  multihyp_data,
  top = 20,
  abrv = 50,
  sizes = TRUE,
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  title = ""
)
}
\arguments{
\item{multihyp_data}{A list of hyp objects}

\item{top}{Limit number of genesets shown}

\item{abrv}{Abbreviation length of genesetlabels}

\item{sizes}{Size dots by geneset sizes}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value e.g. c("fdr", "pval")}

\item{title}{Plot title}
}
\value{
A ggplot object
}
\description{
Plot top enriched genesets across multiple signatures
}
\keyword{internal}
