% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_dots.R
\name{.dots_plot}
\alias{.dots_plot}
\title{Plot top enriched genesets}
\usage{
.dots_plot(hyp_df, title = "", top = 20, abrv = 50, sizes = TRUE,
  val = c("fdr", "pval"))
}
\arguments{
\item{hyp_df}{A dataframe from a hyp object}

\item{title}{Plot title}

\item{top}{Limit number of genesets shown}

\item{abrv}{Abbreviation length of genesetlabels}

\item{sizes}{Size dots by geneset sizes}

\item{val}{Choose significance value e.g. c("fdr", "pval")}
}
\value{
A ggplot object
}
\description{
Plot top enriched genesets
}
\keyword{internal}
