% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{.enrichment_map}
\alias{.enrichment_map}
\title{Plot enrichment map}
\usage{
.enrichment_map(hyp_df, gsets, title,
  similarity_metric = c("jaccard_similarity", "overlap_similarity"),
  similarity_cutoff = 0.2, pval_cutoff = 1, fdr_cutoff = 1,
  val = c("fdr", "pval"), top = NULL)
}
\arguments{
\item{hyp_df}{A dataframe from a hyp object}

\item{gsets}{A list of genesets}

\item{title}{Plot title}

\item{similarity_metric}{Metric to calculate geneset similarity}

\item{similarity_cutoff}{Geneset similarity cutoff}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value shown above nodes e.g. c("fdr", "pval")}

\item{top}{Limit number of pathways shown}
}
\value{
A visNetwork object
}
\description{
Plot enrichment map
}
\examples{
# Grab a list of curated gene sets
gsets <- readRDS(system.file("extdata/gsets.rds", package="hypeR"))
REACTOME <- gsets$REACTOME

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(symbols, REACTOME, bg=2522, fdr=0.05)

# Visualize
hyp_emap(hyp_obj, top=30, val="fdr")

}
\keyword{internal}
