% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_plot.R
\name{.enrichment_plot}
\alias{.enrichment_plot}
\title{Plot top enriched pathways}
\usage{
.enrichment_plot(hyp_df, title, top = 10, val = c("fdr", "pval"))
}
\arguments{
\item{hyp_df}{A dataframe from a hyp object}

\item{title}{Plot title}

\item{top}{Limit number of pathways shown}

\item{val}{Choose significance value e.g. c("fdr", "pval")}
}
\value{
A plotly object
}
\description{
Plot top enriched pathways
}
\keyword{internal}
