% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_members}
\alias{.find_members}
\title{Find geneset members}
\usage{
.find_members(id, genesets, nodes, edges)
}
\arguments{
\item{id}{A vector of ids}

\item{genesets}{A list of genesets (see \code{rgsets})}

\item{nodes}{A data frame of labeled nodes (see \code{rgsets})}

\item{edges}{A data frame of directed edges (see \code{rgsets})}
}
\value{
A vector of ids
}
\description{
Find geneset members
}
\keyword{internal}
