% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_str}
\alias{.format_str}
\title{Format a string using placeholders}
\usage{
.format_str(string, ...)
}
\arguments{
\item{string}{A an unformatted string with placeholders}

\item{...}{Variables to format placeholders with}
}
\value{
A formatted string
}
\description{
Format a string using placeholders
}
\examples{
\dontrun{
format_str("Format with {1} and {2}", "x", "y")
}

}
\keyword{internal}
