% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hyper.R
\name{.github_rds}
\alias{.github_rds}
\title{Load an rds file directly from github}
\usage{
.github_rds(url)
}
\arguments{
\item{url}{A url}
}
\value{
A list
}
\description{
Load an rds file directly from github
}
\keyword{internal}
