% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.hexa}
\alias{.hexa}
\title{Adjust alpha of a hex string}
\usage{
.hexa(hex, percent = 1)
}
\arguments{
\item{hex}{A 6-character hex string (e.g. #000000)}

\item{percent}{Alpha level from 0-1}
}
\value{
A hex string
}
\description{
Adjust alpha of a hex string
}
\keyword{internal}
