% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_hmap.R
\name{.hiearchy_map}
\alias{.hiearchy_map}
\title{Plot hiearchy map}
\usage{
.hiearchy_map(
  hyp_df,
  rgsets_obj,
  pval_cutoff = 1,
  fdr_cutoff = 1,
  val = c("fdr", "pval"),
  top = NULL,
  title = ""
)
}
\arguments{
\item{hyp_df}{A dataframe from a hyp object}

\item{rgsets_obj}{A relatonal geneset from a hyp object}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value displayed when hovering nodes e.g. c("fdr", "pval")}

\item{top}{Limit number of pathways shown}

\item{title}{Plot title}
}
\value{
A visNetwork object
}
\description{
Plot hiearchy map
}
\keyword{internal}
