% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.jaccard_similarity}
\alias{.jaccard_similarity}
\title{Calculate jaccard similarity of two sets}
\usage{
.jaccard_similarity(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
A numerical value
}
\description{
Calculate jaccard similarity of two sets
}
\keyword{internal}
