% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_enrichment.R
\name{.kstest}
\alias{.kstest}
\title{One-sided Kolmogorov–Smirnov test}
\usage{
.kstest(
  n.x,
  y,
  weights = NULL,
  weights.pwr = 1,
  absolute = FALSE,
  plotting = FALSE,
  plot.title = ""
)
}
\arguments{
\item{n.x}{The length of a ranked list}

\item{y}{A vector of positions in the ranked list}

\item{weights}{Weights for weighted score (Subramanian et al.)}

\item{weights.pwr}{Exponent for weights (Subramanian et al.)}

\item{absolute}{Takes max-min score rather than the max deviation from null}

\item{plotting}{Use true to generate plot}

\item{plot.title}{Plot title}
}
\value{
A list of data and plots
}
\description{
One-sided Kolmogorov–Smirnov test
}
\keyword{internal}
