% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.overlap_similarity}
\alias{.overlap_similarity}
\title{Calculate overlap similarity of two sets}
\usage{
.overlap_similarity(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
A numerical value
}
\description{
Calculate overlap similarity of two sets
}
\keyword{internal}
