% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.reverselog_trans}
\alias{.reverselog_trans}
\title{Custom reverse log transformation of continous ggplot axes}
\usage{
.reverselog_trans(base = exp(1))
}
\arguments{
\item{base}{Logarithm base}
}
\description{
Custom reverse log transformation of continous ggplot axes
}
\keyword{internal}
