% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download.gsets}
\alias{download.gsets}
\title{Fetch a gene sets from msigdb}
\usage{
download.gsets(species, category, subcategory)
}
\arguments{
\item{species}{A species to determine gene symbols}

\item{category}{Gene set category}

\item{subcategory}{Gene set subcategory}
}
\value{
A list of gene sets
}
\description{
Fetch a gene sets from msigdb
}
