% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download.msigdb}
\alias{download.msigdb}
\title{Update available gene sets from msigdb}
\usage{
download.msigdb(species = "Homo sapiens")
}
\arguments{
\item{species}{A species to determine gene symbols}
}
\value{
None
}
\description{
Update available gene sets from msigdb
}
