% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_available}
\alias{enrichr_available}
\title{Get enrichr available genesets}
\usage{
enrichr_available()
}
\value{
A dataframe of available genesets
}
\description{
Get enrichr available genesets
}
\examples{
enrichr_available()

}
