% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_download}
\alias{enrichr_download}
\title{Download data from enrichr in the form of a named list}
\usage{
enrichr_download(genesets)
}
\arguments{
\item{genesets}{A name corresponding to available genesets}
}
\value{
A list of genesets
}
\description{
Download data from enrichr in the form of a named list
}
\examples{
ATLAS <- enrichr_download("Human_Gene_Atlas")

}
