% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_gsets}
\alias{enrichr_gsets}
\title{Download data from enrichr in the form of a gsets object}
\usage{
enrichr_gsets(
  genesets,
  db = c("Enrichr", "YeastEnrichr", "FlyEnrichr", "WormEnrichr", "FishEnrichr"),
  clean = FALSE
)
}
\arguments{
\item{genesets}{A name corresponding to available genesets}

\item{db}{A species}

\item{clean}{Use true to clean labels of genesets}
}
\value{
A gsets object
}
\description{
Download data from enrichr in the form of a gsets object
}
\examples{
ATLAS <- enrichr_gsets("Human_Gene_Atlas")

}
