% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_urls}
\alias{enrichr_urls}
\title{Get url base for species-specific enrichr libraries}
\usage{
enrichr_urls(
  db = c("Enrichr", "YeastEnrichr", "FlyEnrichr", "WormEnrichr", "FishEnrichr")
)
}
\arguments{
\item{db}{A species}
}
\value{
A url
}
\description{
Get url base for species-specific enrichr libraries
}
\keyword{internal}
