% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{ex_get}
\alias{ex_get}
\title{Fetch example gene sets}
\usage{
ex_get(symbol = c("C2.CP.REACTOME", "C2.CP.BIOCARTA", "C2.CP.KEGG"))
}
\arguments{
\item{symbol}{A symbol corresponding to an example gene set}
}
\value{
A list of gene sets
}
\description{
Fetch example gene sets
}
\examples{
REACTOME <- ex_get("C2.CP.REACTOME")

}
