% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_to_rmd.R
\name{format_str}
\alias{format_str}
\title{Format a string using placeholders}
\usage{
format_str(string, ...)
}
\arguments{
\item{string}{A an unformatted string with placeholders (e.g. "Format with {1} and {2}")}

\item{...}{Variables to format placeholders with}
}
\value{
A formatted string
}
\description{
Format a string using placeholders
}
