% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_set.R
\docType{class}
\name{GeneSet-class}
\alias{GeneSet-class}
\alias{GeneSet}
\title{GeneSet}
\description{
Class \code{GeneSet} represents a compendium (list) of genesets, each representing a set of gene identifiers
}
\section{Slots}{

\describe{
\item{\code{source.file}}{the file (with path) where the geneset originates from}

\item{\code{geneset}}{the list of character vectors each representing a set of gene IDs}

\item{\code{type}}{type of gene ID (gene symbol, ensemble ID, entrez ID, etc.)}

\item{\code{verbose}}{verbose logging of class operations}

\item{\code{do.save}}{undefined}

\item{\code{name}}{of geneset compendium (most often the source.file w/o the path)}
}}

