% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{genesets_Server}
\alias{genesets_Server}
\title{Shiny server module for geneset selection}
\usage{
genesets_Server(id, clean = FALSE)
}
\arguments{
\item{id}{A unique namespace identifier matching to interface}

\item{clean}{Use true to clean geneset names}
}
\value{
Shiny server code
}
\description{
Shiny server module for geneset selection
}
