% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_modules.R
\name{genesets_UI}
\alias{genesets_UI}
\title{Shiny interface module for geneset selection}
\usage{
genesets_UI(id)
}
\arguments{
\item{id}{A unique namespace identifier}
}
\value{
Shiny ui elements
}
\description{
Shiny interface module for geneset selection
}
