% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_.R
\name{h_}
\alias{h_}
\alias{h1_}
\alias{h2_}
\alias{h3_}
\alias{h4_}
\alias{h5_}
\alias{h6_}
\title{Create headings}
\usage{
h_(..., id = NULL, class = NULL, global = NULL, level = 1)

h1_(..., id = NULL, class = NULL, global = NULL)

h2_(..., id = NULL, class = NULL, global = NULL)

h3_(..., id = NULL, class = NULL, global = NULL)

h4_(..., id = NULL, class = NULL, global = NULL)

h5_(..., id = NULL, class = NULL, global = NULL)

h6_(..., id = NULL, class = NULL, global = NULL)
}
\arguments{
\item{...}{a collection of HTML objects, textual
elements (as bare text or Markdown-styled text),
arbitrary attributes (given as named variables).}

\item{id}{a global attribute for specifying a
unique ID for an HTML element (the value must
be unique within the HTML document, though, this
is not currently enforced here). The primary use
of the \code{id} attribute is to point to a style
in a style sheet. The \code{id} is also used by
JavaScript (via the HTML DOM) to manipulate the
element with the specific \code{id}.

When constructing an \code{id}, there are a few
other things to note: (1) it must contain at
least one character, (2) it must not contain any
space characters, and (3) in HTML, all values are
case-insensitive.}

\item{class}{a global attribute for specifying
one or more classnames for an element. The
\code{class} attribute is primarily used in
pointing to a class in a style sheet. However,
this attribute can also be used by a JavaScript
(via the HTML DOM) to make changes to HTML
elements with a specified class. If providing
several classes, one can either use
\code{c("[classname1]", "[classname2]", ...)} or
a single string with space-separated class names.

When constructing class names, there are two
things to keep in mind: (1) they begin with a
letter (\code{A-Z} or \code{a-z}), and (2)
subsequent characters can be letters
(\code{A-Za-z}), digits (\code{0-9}), hyphens,
and underscores.}

\item{global}{provides an opportunity to supply
global attributes other than the \code{id} and
\code{class} attributes (which have their own
arguments). This is most easily accomplished via
the namesake \code{global()} function since it
allows for inline help and validation for each
of the global attributes. For example, setting
the global attributes \code{title} and
\code{lang} can be done by using this:
\code{global = global(lang = "en", title = "my_title")}.}

\item{level}{the level of heading to use for
the \code{h_()} function. Allowed levels are from
\code{1} to \code{6}. This argument is not required
and not available when using the related \code{h1_()},
\code{h2_()}, \code{h3_()}, \code{h4_()},
\code{h5_()}, or \code{h6_()} functions.}
}
\description{
Allows for the creation of an HTML headings
(e.g., \code{<h1>...</h1>}) with whichever content
is nested within.
}
