% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_html_open.R
\name{h_html_open}
\alias{h_html_open}
\title{Create an opening HTML tag}
\usage{
h_html_open(...)
}
\description{
Create an opening HTML tag
either in standalone mode or as part of
an HTML-build pipeline.
}
