% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\docType{class}
\name{hyp-class}
\alias{hyp-class}
\title{An S4 class to represent a hyp object.}
\description{
An S4 class to represent a hyp object.
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe of enrichment data generated by hypeR}

\item{\code{args}}{A list of arguments passed to hypeR}
}}

