% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.R
\docType{data}
\name{hyp}
\alias{hyp}
\title{A hyp object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
hyp
}
\description{
A hyp object
}
\section{Arguments}{

\describe{
  \item{data}{A dataframe returned by \code{hypeR}}
  \item{args}{A list of arguments passed to \code{hypeR}}
}
}

\section{Methods}{


\code{print(hyp)} shows some information about the object data as
well as the arguments used in creating it.

\code{hyp$as.data.frame()} returns the dataframe slot.
}

\section{See Also}{


\code{multihyp}
}

\keyword{datasets}
