% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_hmap.R
\name{hyp_hmap}
\alias{hyp_hmap}
\title{Visualize hyp/multihyp objects as a hiearchy map}
\usage{
hyp_hmap(hyp_obj, title = "", pval_cutoff = 1, fdr_cutoff = 1,
  val = c("fdr", "pval"), top = NULL, multihyp_titles = TRUE,
  show_plots = TRUE, return_plots = FALSE)
}
\arguments{
\item{hyp_obj}{A hyp or multihyp object}

\item{title}{Plot title}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value displayed when hovering nodes e.g. c("fdr", "pval")}

\item{top}{Limit number of pathways shown}

\item{multihyp_titles}{Use false to disable plot titles for multihyp objects}

\item{show_plots}{Use true to show plots}

\item{return_plots}{Use true to return plots}
}
\value{
A visNetwork object or list of visNetwork objects
}
\description{
Visualize hyp/multihyp objects as a hiearchy map
}
\examples{
# Grab a list of curated gene sets
rgsets_lst <- readRDS(system.file("extdata/rgsets.rds", package="hypeR"))
rgsets_obj <- rgsets_lst$REACTOME

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(symbols, rgsets_obj, gsets_relational=TRUE, bg=7842, fdr=0.05)

# Visualize
hyp_hmap(hyp_obj, top=60)

}
