% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hyp_plot}
\alias{hyp_plot}
\title{Visualize top enriched pathways}
\usage{
hyp_plot(df, top = 10, val = c("fdr", "pval"))
}
\arguments{
\item{df}{A hyper dataframe}

\item{top}{Limit number of pathways shown}

\item{val}{Choose significance value e.g. c("pval", "fdr")}
}
\value{
A plotly object
}
\description{
Visualize top enriched pathways
}
\examples{
# Grab a list of curated gene sets
REACTOME <- ex_get("C2.CP.REACTOME")

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp <- hypeR(symbols, REACTOME, bg=2522, fdr=0.05)

# Visualize
hyp_plot(hyp, top=3, val="fdr")

}
