% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_show.R
\name{hyp_show}
\alias{hyp_show}
\title{Convert hyp object to an interactive datatable}
\usage{
hyp_show(hyp_obj, simple = FALSE, stylish = FALSE)
}
\arguments{
\item{hyp_obj}{A hyp object}

\item{simple}{Use true to only include essential dataframe columns}

\item{stylish}{Use true to add a bootstrap styling theme to datatable}
}
\value{
A datatable object
}
\description{
Convert hyp object to an interactive datatable
}
\examples{
gsets <- readRDS(file.path(system.file("extdata", package="hypeR"), "hypdat.rds"))$gsets

signature <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
               "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
               "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(signature, gsets, bg=2522, fdr_cutoff=0.05)

# Export
hyp_show(hyp_obj)

}
