% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_test.R
\name{hyp_test}
\alias{hyp_test}
\title{Prints the p-value and a recommendation for rejecting or accepting the null hypothesis.}
\usage{
hyp_test(data, mean_0, alpha)
}
\arguments{
\item{data}{A 1-dimensional array. Non-numerical values result in an error}

\item{mean_0}{The mean (as a number) of the population under the null hypothesis. Default: 0.}

\item{alpha}{The threshold for Type I errors as an float.}
}
\value{
The p-value
}
\description{
Prints the p-value and a recommendation for rejecting or accepting the null hypothesis.
}
