% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_to_report.R
\name{hyp_to_report}
\alias{hyp_to_report}
\title{Export hyp object to rmarkdown}
\usage{
hyp_to_report(
  obj,
  file_path,
  title = "Workflow Report",
  subtitle = "",
  author = "",
  show_dots = TRUE,
  show_emaps = TRUE,
  show_hmaps = FALSE,
  hyp_dots_args = list(top = 15, val = "fdr"),
  hyp_emap_args = list(top = 25, val = "fdr", similarity_metric = "jaccard_similarity",
    similarity_cutoff = 0.2),
  hyp_hmap_args = list(top = 25, val = "fdr"),
  session_info = TRUE
)
}
\description{
Export hyp object to rmarkdown
}
\keyword{internal}
