% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hypdat}
\alias{hypdat}
\title{A list of example data for hypeR analysis}
\format{A list of 3 variables:
\describe{
  \item{eset}{An expression set}
  \item{limma}{Differential expression analysis}
  \item{wgcna}{Weighted gene co-expression analysis}
}}
\usage{
hypdat
}
\description{
A list of example data for hypeR analysis
}
\keyword{datasets}
