% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_enrichment.R
\name{hyper.enrichment}
\alias{hyper.enrichment}
\title{Carry out set enrichment test based on hyper-geometric distribution}
\usage{
hyper.enrichment(drawn, categories,
  ntotal = length(unique(unlist(categories))), min.drawsize = 4,
  mht = TRUE, verbose = FALSE)
}
\arguments{
\item{drawn}{One or more sets of 'drawn' items (e.g., genes). Basically, a list of signatures.}

\item{categories}{list of gene sets (e.g., MSigDB c2)}

\item{ntotal}{background population, i.e., the total no. of items from which items are supposed to have been drawn}

\item{min.drawsize}{min no. of drawn items that must be among categories' items}

\item{mht}{correct for multiple hypothesis testing across multiple 'draws'}
}
\value{
a data.frame with rows indexed by the signature(s) tested
}
\description{
Carry out set enrichment test based on hyper-geometric distribution
}
