% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hyper.R
\name{hyperdb_available}
\alias{hyperdb_available}
\title{Check available data to download from hyperdb}
\usage{
hyperdb_available()
}
\description{
Check available data to download from hyperdb
}
\examples{
hyperdb_available()

}
