% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdb.R
\name{hyperdb_fetch}
\alias{hyperdb_fetch}
\title{Fetch gsets from hypeR-db}
\usage{
hyperdb_fetch(type = c("gsets", "rgsets"), gset)
}
\arguments{
\item{type}{Select the type of geneset desired e.g. c("gsets", "rgsets")}

\item{gset}{A name corresponding to an available geneset}
}
\value{
A list of gene sets
}
\description{
Fetch gsets from hypeR-db
}
\examples{
gsets <- hyperdb_fetch(type="gsets", "KEGG_2019_Human")
rgsets <- hyperdb_fetch(type="rgsets", "KEGG")

}
