% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdb.R
\name{hyperdb_get}
\alias{hyperdb_get}
\title{Fetch a gene sets from hyperdb}
\usage{
hyperdb_get(gset)
}
\arguments{
\item{gset}{A name corresponding to an available gene set}
}
\value{
A list of gene sets
}
\description{
Fetch a gene sets from hyperdb
}
