% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hyper.R
\name{hyperdb_gsets}
\alias{hyperdb_gsets}
\title{Download data from hyperdb}
\usage{
hyperdb_gsets(source, gsets)
}
\arguments{
\item{source}{A source identifier}

\item{gsets}{A genesets identifier}
}
\value{
A list
}
\description{
Download data from hyperdb
}
\examples{
KEGG <- hyperdb_gsets("KEGG", "KEGG_v92.0.rds")

}
