% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdb.R
\name{hyperdb_info}
\alias{hyperdb_info}
\title{Print hypeR-db gsets}
\usage{
hyperdb_info(quiet = FALSE)
}
\arguments{
\item{quiet}{Use true to suppress printing of available genesets}
}
\value{
A character vector of available genesets
}
\description{
Print hypeR-db gsets
}
\examples{
hyperdb_info()

}
