% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hyper.R
\name{hyperdb_rgsets}
\alias{hyperdb_rgsets}
\title{Download data from hyperdb in the form of a rgsets object}
\usage{
hyperdb_rgsets(rgsets, version)
}
\arguments{
\item{rgsets}{A name corresponding to an available relational genesets object}

\item{version}{A version number}
}
\value{
An rgsets object
}
\description{
Download data from hyperdb in the form of a rgsets object
}
\examples{
REACTOME <- hyperdb_rgsets("REACTOME", "70.0")

}
