% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{limma}
\alias{limma}
\title{Differential Expression}
\format{
An object of class \code{data.frame} with 9682 rows and 5 columns.
}
\usage{
limma
}
\description{
A differential expression table
}
\keyword{datasets}
