% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_available}
\alias{msigdb_available}
\title{Get msigdbr available genesets}
\usage{
msigdb_available(species = "Homo sapiens")
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}
}
\value{
A dataframe of available genesets
}
\description{
Get msigdbr available genesets
}
\examples{
msigdb_available("Homo sapiens")

}
