% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_check_species}
\alias{msigdb_check_species}
\title{Check if species is available}
\usage{
msigdb_check_species(species = "")
}
\arguments{
\item{species}{A species}
}
\description{
Check if species is available
}
\examples{
\dontrun{
msigdb_check_species("Homo sapiens")
}

}
\keyword{internal}
