% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{msigdb_download_all}
\alias{msigdb_download_all}
\title{Download all gsets from msigdb for a species}
\usage{
msigdb_download_all(species = "Homo sapiens", output_dir = NULL)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{output_dir}{A directory path where gene sets are downloaded instead of to the package location}
}
\value{
A list containing the output directory and version number of gene sets
}
\description{
Download all gsets from msigdb for a species
}
\examples{
msigdb_path <- msigdb_download_all("Homo sapiens")

}
