% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{msigdb_helper}
\alias{msigdb_helper}
\title{Fetch a gene sets from msigdb}
\usage{
msigdb_helper(species = "Homo sapiens", category, subcategory)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{category}{Gene set category (refer to ?msigdbr::msigdbr for avilable categories)}

\item{subcategory}{Gene set subcategory (refer to ?msigdbr::msigdbr for avilable subcategories)}
}
\value{
A list of gene sets
}
\description{
Fetch a gene sets from msigdb
}
\examples{
HALLMARK <- msigdb_helper("Homo sapiens", "H", "")

}
