% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_species}
\alias{msigdb_species}
\title{Get msigdbr available species}
\usage{
msigdb_species()
}
\value{
A character vector of species
}
\description{
Get msigdbr available species
}
\examples{
msigdb_species()

}
