% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_version}
\alias{msigdb_version}
\title{Get msigdbr package version number}
\usage{
msigdb_version()
}
\value{
Version number
}
\description{
Get msigdbr package version number
}
\examples{
msigdb_version()

}
