% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyp.R
\name{multihyp}
\alias{multihyp}
\title{A multihyp object}
\description{
A multihyp object

A multihyp object
}
\section{See Also}{


\code{hyp}
}

\examples{
data <- data.frame(replicate(5,sample(0:1,10,rep=TRUE)))
args <- list("arg_1"=1, "arg_2"=2, "arg_3"=3)
hyp_obj <- hyp$new(data, args=args)
data <- list("hyp_1"=hyp_obj, "hyp_2"=hyp_obj,"hyp_3"=hyp_obj)
multihyp_obj <- multihyp$new(data)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A list of hyp objects}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{multihyp$new()}}
\item \href{#method-print}{\code{multihyp$print()}}
\item \href{#method-as.list}{\code{multihyp$as.list()}}
\item \href{#method-clone}{\code{multihyp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Create a multihyp object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multihyp$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A list of hyp objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new multihyp object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
Print multihyp obect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multihyp$print(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.list"></a>}}
\subsection{Method \code{as.list()}}{
Print multihyp obect
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multihyp$as.list(...)}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of hyp objects as dataframes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{multihyp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
