% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{overlap_similarity}
\alias{overlap_similarity}
\title{Calculate overlap similarity of two sets}
\usage{
overlap_similarity(a, b)
}
\description{
Calculate overlap similarity of two sets
}
\keyword{internal}
