% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rctbl_build}
\alias{rctbl_build}
\title{Reactable builder for hyp or mhyp objects}
\usage{
rctbl_build(obj, ...)
}
\arguments{
\item{obj}{A hyp or multihyp object}

\item{...}{Arguments passed to table generators}
}
\description{
Reactable builder for hyp or mhyp objects
}
