% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{rctbl_mhyp}
\alias{rctbl_mhyp}
\title{Reactable table for multihyp objects}
\usage{
rctbl_mhyp(
  mhyp,
  show_emaps = FALSE,
  show_hmaps = FALSE,
  hyp_emap_args = list(top = 25, val = "fdr"),
  hyp_hmap_args = list(top = 25, val = "fdr")
)
}
\arguments{
\item{mhyp}{A multihyp object}

\item{show_emaps}{Option to show enrichment maps in tabs}

\item{show_hmaps}{Option to show hiearchy maps in tabs}

\item{hyp_emap_args}{A list of keyword arguments passed to hyp_emap}

\item{hyp_hmap_args}{A list of keyword arguments passed to hyp_hmap}
}
\description{
Reactable table for multihyp objects
}
\keyword{export}
