% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_in_viewer.R
\name{show_in_viewer}
\alias{show_in_viewer}
\title{Show constructed HTML content in the
RStudio Viewer pane}
\usage{
show_in_viewer(x)
}
\description{
View the state of the HTML object
in the RStudio Viewer. Stepwise
constructor functions are used to
build the page.
}
