% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatures}
\alias{signatures}
\title{A list of hypeR-compatible gene signatures}
\format{A list of 3 variables:
\describe{
  \item{signature}{A charcter vector of gene symbols}
  \item{experiment}{A list of signatures}
  \item{project}{A list of experiments}
}}
\usage{
signatures
}
\description{
Different ways of passing gene signatures to hypeR and related functions
}
\keyword{datasets}
