% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_page.R
\name{start_page}
\alias{start_page}
\title{Create a page object}
\usage{
start_page()
}
\description{
Create an HTML page object, which can
be passed to other builder functions.
This starts a page with the
necessary components.
}
\examples{
# Start building an HTML page with the
# `start_page()` function
page <- start_page()

# The page is a simple list object
page

# The page can be viewed in RStudio
# with the `show_in_viewer()` function,
# but, let's add some viewable content
# first with `p_()`:
page \%>\%
  p_("**Hello!** This is just text.") \%>\%
  show_in_viewer()
}
