% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_to_rmd.R
\name{string_args}
\alias{string_args}
\title{Convert an arguments list to string format}
\usage{
string_args(args)
}
\arguments{
\item{args}{A list of keyword arguments}
}
\value{
A string of keyword arguments
}
\description{
Convert an arguments list to string format
}
\examples{
string_args(list(x=15, y="fdr" z=TRUE))

}
\keyword{internal}
