% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test.R
\name{t_test}
\alias{t_test}
\title{Calculates a t-statistic of the data relative to a mean which may be specified.}
\usage{
t_test(data, mean_0 = 0)
}
\arguments{
\item{data}{A 1D vector of numbers. Non-numerical values result in an error.}

\item{mean_0}{A 1D vector of numbers. Non-numerical values result in an error.}
}
\value{
The t statistic
}
\description{
Calculates a t-statistic of the data relative to a mean which may be specified.
}
