% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_enrichment.R
\name{VERBOSE}
\alias{VERBOSE}
\title{Function wrapper to control verbosity}
\usage{
VERBOSE(v, ...)
}
\arguments{
\item{v}{A function}

\item{...}{Extra arguments}
}
\value{
A function
}
\description{
Function wrapper to control verbosity
}
