% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{db_get}
\alias{db_get}
\title{Fetch downloaded gene sets from msigdb}
\usage{
db_get(msigdb_info, symbol = c("C1", "C2.CGP", "C2.CP", "C2.CP.BIOCARTA",
  "C2.CP.KEGG", "C2.CP.REACTOME", "C3.MIR", "C3.TFT", "C4.CGN", "C4.CM",
  "C5.BP", "C5.CC", "C5.MF", "C6", "C7", "H"))
}
\arguments{
\item{msigdb_info}{A list containing the gene set directory and version number of gene sets}

\item{symbol}{A symbol corresponding to a msigdb gene set}
}
\value{
A list of gene sets
}
\description{
Fetch downloaded gene sets from msigdb
}
\examples{
msigdb_info <- download_msigdb()
REACTOME <- db_get(msigdb_info, "C2.CP.REACTOME")

}
