% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{.enrichment_map}
\alias{.enrichment_map}
\title{Plot enrichment map}
\usage{
.enrichment_map(df, gsets, title,
  similarity_metric = c("jaccard_similarity", "overlap_similarity"),
  similarity_cutoff = 0.2, pval_cutoff = 1, fdr_cutoff = 1,
  val = c("fdr", "pval"), top = NULL)
}
\description{
Plot enrichment map
}
