% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{.enrichment_map}
\alias{.enrichment_map}
\title{Plot enrichment map}
\usage{
.enrichment_map(df, gsets, title,
  similarity_metric = c("jaccard_similarity", "overlap_similarity"),
  similarity_cutoff = 0.2, pval_cutoff = 1, fdr_cutoff = 1,
  val = c("fdr", "pval"), top = NULL)
}
\arguments{
\item{df}{A dataframe}

\item{gsets}{A list of genesets}

\item{title}{Plot title}

\item{similarity_metric}{Metric to calculate geneset similarity}

\item{similarity_cutoff}{Geneset similarity cutoff}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value shown above nodes e.g. c("fdr", "pval")}

\item{top}{Limit number of pathways shown}
}
\value{
A plotly object
}
\description{
Plot enrichment map
}
