% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyper_enrichment.R
\name{.hyper_enrichment}
\alias{.hyper_enrichment}
\title{Overrepresentation test via hyper-geometric distribution}
\usage{
.hyper_enrichment(
  signature,
  genesets,
  background = length(unique(unlist(genesets))),
  plotting = TRUE
)
}
\arguments{
\item{signature}{A vector of symbols}

\item{genesets}{A list of genesets}

\item{background}{Size of background population genes}

\item{plotting}{Use true to generate plots}
}
\value{
A list of data and plots
}
\description{
Overrepresentation test via hyper-geometric distribution
}
\keyword{internal}
