% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_gsets}
\alias{download_gsets}
\title{Fetch a gene sets from msigdb}
\usage{
download_gsets(species = "Homo sapiens", category, subcategory)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{category}{Gene set category (refer to ?msigdbr::msigdbr for avilable categories)}

\item{subcategory}{Gene set subcategory (refer to ?msigdbr::msigdbr for avilable subcategories)}
}
\value{
A list of gene sets
}
\description{
Fetch a gene sets from msigdb
}
\examples{
HALLMARK <- download_gsets("Homo sapiens", "H", "")

}
