% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{download_msigdb}
\alias{download_msigdb}
\title{Update available gene sets from msigdb}
\usage{
download_msigdb(species = "Homo sapiens", output_dir = NULL)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to the msigdbr R package for available species)}

\item{output_dir}{A directory path where gene sets are downloaded instead of to the package location}
}
\value{
None
}
\description{
Update available gene sets from msigdb
}
