% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_connect}
\alias{enrichr_connect}
\title{Connect to the enrichr web application}
\usage{
enrichr_connect(endpoint)
}
\arguments{
\item{endpoint}{The url endpoint to connect to}
}
\value{
A web response
}
\description{
Connect to the enrichr web application
}
\keyword{internal}
