% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_enrichr.R
\name{enrichr_gsets}
\alias{enrichr_gsets}
\title{Download data from enrichr in the form of a gsets object}
\usage{
enrichr_gsets(genesets)
}
\arguments{
\item{genesets}{A name corresponding to available genesets}
}
\value{
A gsets object
}
\description{
Download data from enrichr in the form of a gsets object
}
\examples{
ATLAS <- enrichr_gsets("Human_Gene_Atlas")

}
