% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsets.R
\docType{data}
\name{gsets}
\alias{gsets}
\title{A genesets object}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
gsets
}
\description{
A genesets object
}
\section{Arguments}{

\describe{
  \item{genesets}{A named list of genesets}
  \item{name}{A character vector describing source of genesets}
  \item{version}{A character vector describing versioning}
}
}

\section{Methods}{


\code{print(gsets)} prints geneset information.

\code{gsets$reduce(background)} reduces genesets to a background distribution of symbols.
}

\section{See Also}{


\code{rgsets}
}

\examples{
genesets <- list("GSET1" = c("GENE1", "GENE2", "GENE3"),
                 "GSET2" = c("GENE4", "GENE5", "GENE6"),
                 "GSET3" = c("GENE7", "GENE8", "GENE9"))

gsets_obj <- gsets$new(genesets, name="example", version="v1.0")
print(gsets_obj)

}
\keyword{datasets}
