% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsets.R
\name{gsets}
\alias{gsets}
\title{A genesets object}
\description{
A genesets object

A genesets object
}
\section{Arguments}{

\describe{
  \item{genesets}{A named list of genesets}
  \item{name}{A character vector describing source of genesets}
  \item{version}{A character vector describing versioning}
}
}

\section{Methods}{


\code{print(gsets)} prints geneset information.

\code{gsets$reduce(background)} reduces genesets to a background distribution of symbols.
}

\section{See Also}{


\code{rgsets}
}

\examples{
genesets <- list("GSET1" = c("GENE1", "GENE2", "GENE3"),
                 "GSET2" = c("GENE4", "GENE5", "GENE6"),
                 "GSET3" = c("GENE7", "GENE8", "GENE9"))

gsets_obj <- gsets$new(genesets, name="example", version="v1.0")
print(gsets_obj)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{gsets$new()}}
\item \href{#method-print}{\code{gsets$print()}}
\item \href{#method-reduce}{\code{gsets$reduce()}}
\item \href{#method-clone}{\code{gsets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gsets$new(genesets, name = "", version = "", quiet = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gsets$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce"></a>}}
\subsection{Method \code{reduce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gsets$reduce(background)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{gsets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
