% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_.R
\name{h_}
\alias{h_}
\alias{h1_}
\alias{h2_}
\alias{h3_}
\alias{h4_}
\alias{h5_}
\alias{h6_}
\title{Create headings}
\usage{
h_(..., level = 1)

h1_(...)

h2_(...)

h3_(...)

h4_(...)

h5_(...)

h6_(...)
}
\description{
Create H1-6 tags either in
standalone mode or as part of an
HTML-build pipeline.
}
