% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\docType{class}
\name{hyp-class}
\alias{hyp-class}
\title{An S4 class to represent a hyper object.}
\description{
An S4 class to represent a hyper object.
}
\section{Slots}{

\describe{
\item{\code{data}}{A dataframe}
}}

