% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.R
\docType{data}
\name{hyp}
\alias{hyp}
\title{A hyp object}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
hyp
}
\description{
A hyp object
}
\section{Arguments}{

\describe{
  \item{data}{A dataframe returned by \code{hypeR}}
  \item{plots}{A list of plots returned by \code{hypeR}}
  \item{args}{A list of arguments passed to \code{hypeR}}
}
}

\section{Methods}{


\code{print(hyp)} shows some information about the object data as
well as the arguments used in creating it.

\code{hyp$as.data.frame()} returns the dataframe slot.
}

\section{See Also}{


\code{multihyp}
}

\examples{
data <- data.frame(replicate(5,sample(0:1,10,rep=TRUE)))
args <- list("arg_1"=1, "arg_2"=2, "arg_3"=3)
hyp_obj <- hyp$new(data, args=args)

}
\keyword{datasets}
