% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp.R
\name{hyp}
\alias{hyp}
\title{A hyp object}
\description{
A hyp object

A hyp object
}
\section{Arguments}{

\describe{
  \item{data}{A dataframe returned by \code{hypeR}}
  \item{plots}{A list of plots returned by \code{hypeR}}
  \item{args}{A list of arguments passed to \code{hypeR}}
}
}

\section{Methods}{


\code{print(hyp)} shows some information about the object data as
well as the arguments used in creating it.

\code{hyp$as.data.frame()} returns the dataframe slot.
}

\section{See Also}{


\code{multihyp}
}

\examples{
data <- data.frame(replicate(5,sample(0:1,10,rep=TRUE)))
args <- list("arg_1"=1, "arg_2"=2, "arg_3"=3)
hyp_obj <- hyp$new(data, args=args)

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hyp$new()}}
\item \href{#method-print}{\code{hyp$print()}}
\item \href{#method-as.data.frame}{\code{hyp$as.data.frame()}}
\item \href{#method-clone}{\code{hyp$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hyp$new(data, plots = NULL, args = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hyp$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-as.data.frame"></a>}}
\subsection{Method \code{as.data.frame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hyp$as.data.frame(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hyp$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
