% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{hyp_emap}
\alias{hyp_emap}
\title{Visualize enrichment map from one or more signatures}
\usage{
hyp_emap(hyp, gsets, similarity_metric = c("jaccard_similarity",
  "overlap_similarity"), similarity_cutoff = 0.2, pval_cutoff = 1,
  fdr_cutoff = 1, val = c("fdr", "pval"), top = NULL,
  show_plots = TRUE, return_plots = FALSE)
}
\description{
Visualize enrichment map from one or more signatures
}
