% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_emap.R
\name{hyp_emap}
\alias{hyp_emap}
\title{Visualize hyp or multihyp objects as an enrichment map}
\usage{
hyp_emap(hyp_obj, title = "",
  similarity_metric = c("jaccard_similarity", "overlap_similarity"),
  similarity_cutoff = 0.2, pval_cutoff = 1, fdr_cutoff = 1,
  val = c("fdr", "pval"), top = NULL, multihyp_titles = TRUE,
  show_plots = TRUE, return_plots = FALSE)
}
\arguments{
\item{hyp_obj}{A hyp or multihyp object}

\item{title}{Plot title}

\item{similarity_metric}{Metric to calculate geneset similarity}

\item{similarity_cutoff}{Geneset similarity cutoff}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{val}{Choose significance value shown above nodes e.g. c("fdr", "pval")}

\item{top}{Limit number of pathways shown}

\item{multihyp_titles}{Use false to disable plot titles for multihyp objects}

\item{show_plots}{An option to show plots}

\item{return_plots}{An option to return plots}
}
\value{
A visNetwork object or list of visNetwork objects
}
\description{
Visualize hyp or multihyp objects as an enrichment map
}
