% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyp_plot.R
\name{hyp_plot}
\alias{hyp_plot}
\title{Visualize hyp or multihyp objects as a bar plot}
\usage{
hyp_plot(hyp_obj, title = "", top = 10, val = c("fdr", "pval"),
  multihyp_titles = TRUE, show_plots = TRUE, return_plots = FALSE)
}
\arguments{
\item{hyp_obj}{A hyp or multihyp object}

\item{title}{Plot title}

\item{top}{Limit number of pathways shown}

\item{val}{Choose significance value e.g. c("fdr", "pval")}

\item{multihyp_titles}{Use false to disable plot titles for multihyp objects}

\item{show_plots}{An option to show plots}

\item{return_plots}{An option to return plots}
}
\value{
A plotly object
}
\description{
Visualize hyp or multihyp objects as a bar plot
}
\examples{
# Grab a list of curated gene sets
gsets <- readRDS(system.file("extdata/gsets.rds", package="hypeR"))
REACTOME <- gsets$REACTOME

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(symbols, REACTOME, bg=2522, fdr=0.05)

# Visualize
hyp_plot(hyp_obj, top=3, val="fdr")

}
