% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hyp_to_excel}
\alias{hyp_to_excel}
\title{Export hyper dataframe to excel}
\usage{
hyp_to_excel(df, file.path, cols = seq_len(ncol(df)))
}
\arguments{
\item{df}{A hyper dataframe}

\item{file.path}{Output file path}

\item{cols}{Dataframe columns to include}
}
\value{
None
}
\description{
Export hyper dataframe to excel
}
\examples{
# Grab a list of curated gene sets
REACTOME <- ex_get("C2.CP.REACTOME")

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp <- hypeR(symbols, REACTOME, bg=2522, fdr=0.05)

# Export
hyp_to_excel(hyp, file.path="pathways.xlsx")

}
