% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hypeR}
\alias{hypeR}
\title{Perform hyper enrichment}
\usage{
hypeR(symbols, gsets, bg = 23467, min.drawsize = 4, pval.cutoff = 1,
  fdr.cutoff = 1, verbose = FALSE)
}
\arguments{
\item{symbols}{A character vector of gene symbols}

\item{gsets}{A list of gene sets}

\item{bg}{Size or character vector of background population genes}

\item{min.drawsize}{Min number of drawn items that must be among categories items}

\item{pval.cutoff}{Filter results to be less than pval cutoff}

\item{fdr.cutoff}{Filter results to be less than fdr cutoff}

\item{verbose}{Use false to suppress logs}
}
\value{
A hyper dataframe
}
\description{
Perform hyper enrichment
}
\examples{
# Grab a list of curated gene sets
REACTOME <- db.get("C2.CP.REACTOME")

# Genes involed in tricarboxylic acid cycle
symbols <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
             "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
             "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp <- hypeR(symbols, REACTOME, bg=2522, fdr=0.05)

}
