% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hypeR}
\alias{hypeR}
\title{Perform hyper enrichment}
\usage{
hypeR(signature, gsets, test = c("hypergeometric", "kstest"),
  bg = 23467, pval_cutoff = 1, fdr_cutoff = 1, weights_pwr = 1,
  absolute = FALSE, do_plots = FALSE, verbose = FALSE)
}
\arguments{
\item{signature}{A vector of symbols}

\item{gsets}{A list of genesets or a relational genesets object}

\item{test}{Choose an enrichment type e.g. c("hypergeometric", "kstest")}

\item{bg}{Size or character vector of background population genes}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{weights_pwr}{Exponent for weights (Subramanian et al.)}

\item{absolute}{Takes max-min score rather than the max deviation from null}

\item{do_plots}{Use true to generate plots}

\item{verbose}{Use false to suppress logs}
}
\value{
A hyp object
}
\description{
Perform hyper enrichment
}
\examples{
gsets <- readRDS(file.path(system.file("extdata", package="hypeR"), "hypdat.rds"))$gsets

signature <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
               "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
               "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(signature, gsets, bg=2522, fdr_cutoff=0.05)

}
