% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\name{hypeR}
\alias{hypeR}
\title{Perform hyper enrichment}
\usage{
hypeR(signature, gsets, test = c("hypergeometric", "kstest"),
  bg = 23467, gsets_relational = FALSE, pval_cutoff = 1,
  fdr_cutoff = 1, weights = NULL, weights_pwr = 1,
  absolute = FALSE, do_plots = FALSE, verbose = FALSE)
}
\arguments{
\item{signature}{A vector of symbols}

\item{gsets}{A list of genesets}

\item{bg}{Size or character vector of background population genes}

\item{gsets_relational}{Use true to inform gsets are relational}

\item{pval_cutoff}{Filter results to be less than pval cutoff}

\item{fdr_cutoff}{Filter results to be less than fdr cutoff}

\item{verbose}{Use false to suppress logs}
}
\value{
A hyp object
}
\description{
Perform hyper enrichment
}
\examples{
# Grab a list of curated gene sets
gsets <- readRDS(system.file("extdata/gsets.rds", package="hypeR"))
REACTOME <- gsets$REACTOME

# Genes involed in tricarboxylic acid cycle
signature <- c("IDH3B","DLST","PCK2","CS","PDHB","PCK1","PDHA1","LOC642502",
               "PDHA2","LOC283398","FH","SDHD","OGDH","SDHB","IDH3A","SDHC",
               "IDH2","IDH1","OGDHL","PC","SDHA","SUCLG1","SUCLA2","SUCLG2")

# Perform hyper enrichment
hyp_obj <- hypeR(signature, REACTOME, bg=2522, fdr_cutoff=0.05)

}
