% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_hyper.R
\name{hyperdb_gsets}
\alias{hyperdb_gsets}
\title{Download data from hyperdb in the form of a gsets object}
\usage{
hyperdb_gsets(gsets, version)
}
\arguments{
\item{gsets}{A name corresponding to an available genesets object}

\item{version}{A version number}
}
\value{
A gsets object
}
\description{
Download data from hyperdb in the form of a gsets object
}
\examples{
SMPDB <- hyperdb_gsets("SMPDB", "2.75")

}
