% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdb.R
\name{hyperdb_info}
\alias{hyperdb_info}
\title{Print available gene sets}
\usage{
hyperdb_info(quiet = FALSE)
}
\arguments{
\item{quiet}{Use true to suppress printing of available genesets}
}
\value{
A character vector of available genesets
}
\description{
Print available gene sets
}
