% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdb.R
\name{hyperdb_info}
\alias{hyperdb_info}
\title{Print hypeR-db gsets}
\usage{
hyperdb_info(type = c("gsets", "rgsets"), quiet = FALSE)
}
\arguments{
\item{type}{Select the type of geneset desired e.g. c("gsets", "rgsets")}

\item{quiet}{Use true to suppress printing of available genesets}
}
\value{
A character vector of available genesets
}
\description{
Print hypeR-db gsets
}
\examples{
available <- hyperdb_info(type="gsets", quiet=TRUE)
available <- hyperdb_info(type="rgsets", quiet=TRUE)

}
