% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msigdb.R
\name{msigdb_download}
\alias{msigdb_download}
\title{Download all gene sets from msigdb for a given species}
\usage{
msigdb_download(species = "Homo sapiens", output_dir = NULL)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{output_dir}{A directory path where gene sets are downloaded instead of to the package location}
}
\value{
A list containing the output directory and version number of gene sets
}
\description{
Download all gene sets from msigdb for a given species
}
\examples{
msigdb_path <- msigdb_download()

}
