% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_download}
\alias{msigdb_download}
\title{Download data from msigdb in the form of a named list}
\usage{
msigdb_download(species, collection, subcollection = NULL)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for available species)}

\item{collection}{Geneset collection (refer to ?msigdbr::msigdbr_collections for available categories)}

\item{subcollection}{Geneset subcollection (refer to ?msigdbr::msigdbr_collections for available subcategories)}
}
\value{
A list of genesets
}
\description{
Download data from msigdb in the form of a named list
}
\examples{
HALLMARK_HUMAN <- msigdb_download("Homo sapiens", "H")
HALLMARK_MOUSE <- msigdb_download("Mus musculus", "MH")

}
