% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_download}
\alias{msigdb_download}
\title{Download data from msigdb in the form of a named list}
\usage{
msigdb_download(species, category, subcategory = "")
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{category}{Geneset category (refer to ?msigdbr::msigdbr for avilable categories)}

\item{subcategory}{Geneset subcategory (refer to ?msigdbr::msigdbr for avilable subcategories)}
}
\value{
A list of genesets
}
\description{
Download data from msigdb in the form of a named list
}
\examples{
HALLMARK <- msigdb_download("Homo sapiens", "H", "")

}
