% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_gsets}
\alias{msigdb_gsets}
\title{Download data from msigdb in the form of a gsets object}
\usage{
msigdb_gsets(species, collection, subcollection = "", clean = FALSE)
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{collection}{Geneset collection (previous argument "category" is deprecated. refer to ?msigdbr::msigdbr for avilable collections)}

\item{subcollection}{Geneset subcollection (previous argument "subcategory" is deprecated. refer to ?msigdbr::msigdbr for avilable subcollections)}

\item{clean}{Use true to clean labels of genesets}
}
\value{
A gsets object
}
\description{
Download data from msigdb in the form of a gsets object
}
\examples{
HALLMARK <- msigdb_gsets("Homo sapiens", "H", "")

}
