% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_msig.R
\name{msigdb_gsets}
\alias{msigdb_gsets}
\title{Download data from msigdb in the form of a gsets object}
\usage{
msigdb_gsets(species, category, subcategory = "")
}
\arguments{
\item{species}{A species to determine gene symbols (refer to ?msigdbr::msigdbr for avilable species)}

\item{category}{Geneset category (refer to ?msigdbr::msigdbr for avilable categories)}

\item{subcategory}{Geneset subcategory (refer to ?msigdbr::msigdbr for avilable subcategories)}
}
\value{
A gsets object
}
\description{
Download data from msigdb in the form of a gsets object
}
\examples{
HALLMARK <- msigdb_gsets("Homo sapiens", "H", "")

}
