% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hype.R
\docType{class}
\name{multihyp-class}
\alias{multihyp-class}
\title{An S4 class to represent multiple hyper obejcts.}
\description{
An S4 class to represent multiple hyper obejcts.
}
\section{Slots}{

\describe{
\item{\code{data}}{A list of hyper objects}
}}

