% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyp.R
\docType{data}
\name{multihyp}
\alias{multihyp}
\title{A multihyp object}
\format{An object of class \code{R6ClassGenerator} of length 25.}
\usage{
multihyp
}
\description{
A multihyp object
}
\section{Arguments}{

\describe{
  \item{data}{A list of \code{hyp} objects}
}
}

\section{Methods}{


\code{print(multihyp)} shows some information about the object data

\code{multihyp$as.list()} returns a list of \code{hyp} objects as dataframes.
}

\section{See Also}{


\code{hyp}
}

\examples{
data <- data.frame(replicate(5,sample(0:1,10,rep=TRUE)))
args <- list("arg_1"=1, "arg_2"=2, "arg_3"=3)
hyp_obj <- hyp$new(data, args=args)
data <- list("hyp_1"=hyp_obj, "hyp_2"=hyp_obj,"hyp_3"=hyp_obj)
multihyp_obj <- multihyp$new(data)

}
\keyword{datasets}
