% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multihyp.R
\docType{data}
\name{multihyp}
\alias{multihyp}
\title{A multihyp object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
multihyp
}
\description{
A multihyp object
}
\section{Arguments}{

\describe{
  \item{data}{A list of \code{hyp} objects}
}
}

\section{Methods}{

\code{multihyp$as.list()} returns a list of \code{hyp} objects as dataframes.
}

\keyword{datasets}
