% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvector.R
\docType{data}
\name{pvector}
\alias{pvector}
\title{A push/pop capable vector}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
pvector
}
\description{
A push/pop capable vector
}
\section{Arguments}{

\describe{
  \item{values}{A vector of values}
}
}

\section{Methods}{


\code{print(pvector)} prints vector values.

\code{pvector$length()} returns length of vector values.
 
\code{pvector$pop()} pops vector, returning popped value.

\code{pvector$push()} pushes to vector.
}

\examples{
pv <- pvector$new(c(1,2,3))
popped <- pv$pop()
pv$push(4)
pv$push(c(5,6))
print(pv) 

}
\keyword{datasets}
