% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvector.R
\name{pvector}
\alias{pvector}
\title{A push/pop capable vector}
\description{
A push/pop capable vector

A push/pop capable vector
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{values}}{A vector of values}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-pvector-new}{\code{pvector$new()}}
\item \href{#method-pvector-print}{\code{pvector$print()}}
\item \href{#method-pvector-length}{\code{pvector$length()}}
\item \href{#method-pvector-pop}{\code{pvector$pop()}}
\item \href{#method-pvector-push}{\code{pvector$push()}}
\item \href{#method-pvector-clone}{\code{pvector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-new"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-new}{}}}
\subsection{Method \code{new()}}{
Create a pvector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$new(values = c())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{values}}{A vector of values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new pvector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-print"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-print}{}}}
\subsection{Method \code{print()}}{
Print pvector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$print()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-length"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-length}{}}}
\subsection{Method \code{length()}}{
Get length of pvector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$length()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-pop"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-pop}{}}}
\subsection{Method \code{pop()}}{
Pop vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$pop()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Popped value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-push"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-push}{}}}
\subsection{Method \code{push()}}{
Push values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$push(pushed.values)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pushed.values}}{A vector of values}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pvector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-pvector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
