% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvector.R
\name{pvector}
\alias{pvector}
\title{A push/pop capable vector}
\description{
A push/pop capable vector

A push/pop capable vector
}
\section{Arguments}{

\describe{
  \item{values}{A vector of values}
}
}

\section{Methods}{


\code{print(pvector)} prints vector values.

\code{pvector$length()} returns length of vector values.
 
\code{pvector$pop()} pops vector, returning popped value.

\code{pvector$push()} pushes to vector.
}

\examples{
pv <- pvector$new(c(1,2,3))
popped <- pv$pop()
pv$push(4)
pv$push(c(5,6))
print(pv) 

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{pvector$new()}}
\item \href{#method-print}{\code{pvector$print()}}
\item \href{#method-length}{\code{pvector$length()}}
\item \href{#method-pop}{\code{pvector$pop()}}
\item \href{#method-push}{\code{pvector$push()}}
\item \href{#method-clone}{\code{pvector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$new(values = c())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-length"></a>}}
\subsection{Method \code{length()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$length()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-pop"></a>}}
\subsection{Method \code{pop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$pop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-push"></a>}}
\subsection{Method \code{push()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$push(pushed.values)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{pvector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
