% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsets.R
\docType{data}
\name{rgsets}
\alias{rgsets}
\title{Relational genes ets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rgsets
}
\description{
Relational genes ets
}
\section{Arguments}{

\describe{
  \item{gsets}{A list of genesets}
  \item{nodes}{A data frame of labeled nodes}
  \item{edges}{A data frame of directed edges}
}
}

\section{Methods}{

}

\keyword{datasets}
