% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsets.R
\docType{data}
\name{rgsets}
\alias{rgsets}
\title{Relational genesets}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rgsets
}
\description{
Relational genesets
}
\section{Arguments}{

\describe{
  \item{gsets}{A list of genesets where list names refers to geneset
  labels and values are geneset members represented as a vector}
  \item{nodes}{A data frame of labeled nodes e.g.
  \cr
\tabular{rrrr}{
  \tab label\cr
  id.1 \tab Geneset Label 1 \tab\cr
  id.2 \tab Geneset Label 2 \tab\cr
  id.3 \tab Geneset Label 3 \tab\cr
}
  }
  \item{edges}{A data frame of directed edges}
}
}

\section{Methods}{

}

\section{See Also}{


\code{pvector}
}

\keyword{datasets}
