% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgsets.R
\name{rgsets}
\alias{rgsets}
\title{A relational genesets object}
\description{
A relational genesets object

A relational genesets object
}
\section{Arguments}{

\describe{
  \item{genesets}{A list of genesets where list names refers to geneset
  labels and values are geneset members represented as a vector}
  \item{nodes}{A data frame of labeled nodes e.g.
    \cr
    \tabular{rrrr}{
      \tab label\cr
      G1 \tab Geneset 1 \tab\cr
      G2 \tab Geneset 2 \tab\cr
      G3 \tab Geneset 3 \tab\cr
    }
  }
  \item{edges}{A data frame of directed edges
    \cr
    \tabular{rrrr}{
      from \tab to\cr
      G1 \tab G2 \tab\cr
      G1 \tab G3 \tab\cr
    }
  }
  \item{name}{A character vector describing source of genesets}
  \item{version}{A character vector describing versioning}
}
}

\section{Methods}{


\code{print(rgsets)} shows some information about the object data

\code{rgsets$subset(labels)} returns an rgsets object subsetted
on geneset labels
}

\section{See Also}{


\code{gsets}
\code{pvector}
}

\examples{
testdat <- readRDS(file.path(system.file("extdata", package="hypeR"), "testdat.rds"))
rgsets <- rgsets$new(genesets=testdat$genesets, nodes=testdat$nodes, edges=testdat$edges, name="Example", version="v1.0")

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{rgsets$new()}}
\item \href{#method-print}{\code{rgsets$print()}}
\item \href{#method-reduce}{\code{rgsets$reduce()}}
\item \href{#method-subset}{\code{rgsets$subset()}}
\item \href{#method-clone}{\code{rgsets$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rgsets$new(genesets, nodes, edges, name = "", version = "", quiet = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rgsets$print(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-reduce"></a>}}
\subsection{Method \code{reduce()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rgsets$reduce(background)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-subset"></a>}}
\subsection{Method \code{subset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rgsets$subset(labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{rgsets$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
