% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_page.R
\name{start_page}
\alias{start_page}
\title{Create a page object}
\usage{
start_page()
}
\description{
Create an HTML page object, which can
be passed to other builder functions.
This starts a page with the
necessary components.
}
